requirements.txt
absl-py==2.0.0
alembic==1.12.1
asttokens==2.4.0
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.3
build==1.0.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
distlib==0.3.7
docopt==0.6.2
executing==1.2.0
Faker==19.12.0
filelock==3.13.0
Flask==3.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.44.0
fsspec==2023.10.0
fuzzysearch==0.7.3
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.1
grpcio==1.59.0
h5py==3.10.0
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.2
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
Mako==1.2.4
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mpmath==1.3.0
mysql-connector-python==8.2.0
mysqlclient==2.2.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.2
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
pip-tools==7.3.0
pipdeptree==2.13.0
pipreqs==0.4.13
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
PuLP==2.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.5.2
Pygments==2.16.1
pyparsing==3.1.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
SQLAlchemy==2.0.22
stack-data==0.6.2
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-cpu==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
torch==2.1.0
torchdata==0.7.0
torchtext==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.24.6
wcwidth==0.2.6
Werkzeug==3.0.1
wrapt==1.14.1
yarg==0.1.9
